use a 2 pass transform for schema and saving AST so we don't need to reparse
