+++ packages/cardhost/app/lib/builder.ts
@@ -9,6 +9,7 @@ import {
   assertDocumentDataIsResource,
   CardModel,
 } from '@cardstack/core/src/interfaces';
+import type { File } from '@babel/types';
 import { RawCardDeserializer } from '@cardstack/core/src/serializers';
 import { fetchJSON } from './jsonapi-fetch';
 import config from 'cardhost/config/environment';
@@ -150,7 +151,7 @@ export default class LocalRealm implements Builder {
       let definedCard = makeGloballyAddressable(
         url,
         compiledCard,
-        (local, type, src) => this.define(url, local, type, src)
+        (local, type, src, ast) => this.define(url, local, type, src, ast)
       );
       this.compiledCardCache.set(url, definedCard);
       return definedCard;
@@ -301,7 +302,8 @@ export default class LocalRealm implements Builder {
     cardURL: string,
     localModule: string,
     type: string,
-    source: string
+    source: string,
+    ast: File | undefined
   ): string {
     let moduleIdentifier = `@cardstack/local-realm-compiled/${encodeCardURL(
       cardURL
@@ -314,7 +316,7 @@ export default class LocalRealm implements Builder {
 
     switch (type) {
       case JS_TYPE: {
-        eval(dynamicCardTransform(moduleIdentifier, source));
+        eval(dynamicCardTransform(moduleIdentifier, source, ast));
         return moduleIdentifier;
       }
       case CSS_TYPE:
